CORE-591: rspace: move Blake2b256Hash to 'rspace' package
