+++ rspace/src/test/scala/coop/rchain/rspace/test/DummyTrieStore.scala
@@ -1,6 +1,7 @@
 package coop.rchain.rspace.test
 
-import coop.rchain.rspace.history.{Blake2b256Hash, ITrieStore, Trie}
+import coop.rchain.rspace.Blake2b256Hash
+import coop.rchain.rspace.history.{ITrieStore, Trie}
 
 class DummyTrieStore[T, K, V] extends ITrieStore[T, K, V] {
